requirements.txt
fastapi
httpx
beautifulsoup4
duckduckgo-search
uvicorn
python-dotenv